*: move upgrade handler to server/handler
